+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ### 0.0.8
 - scrolls & potions with random names
 - location loot
+- statues & altars
+- bushes
+- teleport spell
+
 
 ### 0.0.7
 - init dungeon generator